Update to make package by itself
